move enums to own folder and add MatchResult enum
