+++ DragonFruit.Six.API/AccountInfo.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using DragonFruit.Six.API.Enums;
 using DragonFruit.Six.API.Helpers;
 using DragonFruit.Six.API.Processing;
 using DragonFruit.Six.API.Verification;
@@ -49,9 +50,6 @@ namespace DragonFruit.Six.API
         /// </summary>
         [JsonProperty("userid")]
         public string UbisoftId { get; set; }
-
-        [JsonIgnore]
-        public Verification.Verification AccountStatus => Server.GetUser(Guid);
 
         /// <summary>
         /// Get a user's account info
