+++ DragonFruit.Six.API/Enums/MatchResult.cs
@@ -1,0 +1,13 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+namespace DragonFruit.Six.API.Enums
+{
+    public enum MatchResult
+    {
+        Unknown = 0,
+        Win = 1,
+        Loss = 2,
+        Abandon = 3
+    }
+}
