+++ DragonFruit.Six.API/Enums/Platforms.cs
@@ -1,0 +1,12 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+namespace DragonFruit.Six.API.Enums
+{
+    public enum Platforms
+    {
+        PC = 1,
+        PSN = 2,
+        XB1 = 3
+    }
+}
