+++ DragonFruit.Six.API/Enums/Regions.cs
@@ -1,0 +1,15 @@
+﻿// Dragon6 API Copyright 2020 DragonFruit Network <inbox@dragonfruit.network>
+// Licensed under Apache-2. Please refer to the LICENSE file for more info
+
+using System.Diagnostics.CodeAnalysis;
+
+namespace DragonFruit.Six.API.Enums
+{
+    [SuppressMessage("ReSharper", "InconsistentNaming")]
+    public enum Regions
+    {
+        EMEA,
+        NSCA,
+        APAC
+    }
+}
